Shared CertificateValidation and used in S.N.Quic
